aster_pci::cfg_space

Enum Bar

pub enum Bar {
    Memory(Arc<MemoryBar>),
    Io(Arc<IoBar>),
}

BAR space in PCI common config space.

Variants

Memory(Arc<MemoryBar>)

Memory BAR

Io(Arc<IoBar>)

I/O BAR

Implementations

impl Bar

pub fn read_once<T: PodOnce + PortRead>(&self, offset: usize) -> Result<T>

Reads a value of a specified type at a specified offset.

pub fn write_once<T: PodOnce + PortWrite>( &self, offset: usize, value: T, ) -> Result<()>

Writes a value of a specified type at a specified offset.